Extracted Cart widgets into separate packages Update item count (inside cart) when items quantity change
