+++ client/src/main/java/com/arcbees/beeshop/client/application/widget/sidepanel/cart/cartitem/CartItemView.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.arcbees.beeshop.client.application.widget.sidepanel.cart;
+package com.arcbees.beeshop.client.application.widget.sidepanel.cart.cartitem;
 
 import com.arcbees.beeshop.client.application.ShoppingCartItem;
 import com.arcbees.beeshop.client.resources.AppMessages;
