+++ client/src/main/java/com/arcbees/beeshop/client/application/widget/sidepanel/cart/cartitem/CartItemView.ui.xml
