+++ client/src/main/java/com/arcbees/beeshop/client/application/widget/sidepanel/cart/cartitems/CartItemsModule.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2015 ArcBees Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package com.arcbees.beeshop.client.application.widget.sidepanel.cart.cartitems;
+
+import com.gwtplatform.mvp.client.gin.AbstractPresenterModule;
+
+public class CartItemsModule extends AbstractPresenterModule {
+    @Override
+    protected void configure() {
+        bindPresenterWidget(CartItemsPresenter.class, CartItemsPresenter.MyView.class,
+                CartItemsView.class);
+    }
+}
