+++ client/src/main/java/com/arcbees/beeshop/client/application/widget/sidepanel/cart/cartitems/CartItemsView.ui.xml
